+++ client/src/assets/params.ts
@@ -4,7 +4,7 @@ export const PAD_MOVE_RADIUS_1 = 48;
 export const PAD_FIRE_RADIUS_0 = 16;
 export const PAD_FIRE_RADIUS_1 = 40;
 
-export const BASE_RESOLUTION = 256;
+// 1024x1024 texture, we can up world size or increase map resolution
 export const BOUNDS_SIZE = 0x400;
 
 export const WORLD_SCALE = 60;